refactor: extract public transfer/sweep API into transfer.jl
